requirements.txt
fastapi
uvicorn[standard]
python-multipart
openpyxl
pandas
google-genai
python-dotenv
aiofiles
firebase-admin
pdfplumber
